refactor: split ui.go into focused modules for better maintainability
