+++ internal/ui/core/ui/ui.go
@@ -1,0 +1,9 @@
+package ui
+
+// This file has been split into smaller, focused files:
+// - ui_initialization.go: UI setup and initialization
+// - ui_views.go: View creation and registration
+// - ui_keybindings.go: Key binding handling
+// - ui_navigation.go: View switching and navigation
+// - ui_modals.go: Modal and dialog management
+// - ui_utilities.go: Utility functions and helpers
